string formatting fix for NSInteger

+++ examples/S4/S4/DiscountController.m
@@ -72,7 +72,7 @@
     float realVal = self.discountSlider.value;
     NSInteger intVal = realVal*100.0;
     self.discountLabel.text =
-    [NSString stringWithFormat:@"%ld%%", intVal];
+    [NSString stringWithFormat:@"%ld%%", (long)intVal];
 }
 
 - (IBAction)discountChanged:(id)sender
